feat/sti-service: Move files around, rename DatabaseContext
