+++ Backend/Accounting-Contract/AccountingDatabase.cs
@@ -1,0 +1,8 @@
+using Microsoft.EntityFrameworkCore;
+
+namespace Accounting.Services;
+
+public class AccountingDatabase : DbContext
+{
+    public AccountingDatabase(DbContextOptions<AccountingDatabase> options) : base(options) { }
+}
