--- Backend/Accounting-Contract/DatabaseContext.cs
@@ -1,8 +1,0 @@
-using Microsoft.EntityFrameworkCore;
-
-namespace Accounting.Services;
-
-public class DatabaseContext : DbContext
-{
-    public DatabaseContext(DbContextOptions<DatabaseContext> options) : base(options) { }
-}
